quests: Switching quickstart/meet to have the shorter one be the intro
https://phabricator.endlessm.com/T27928
