+++ eosclubhouse/quests/hack2/quickstart.py
@@ -1,12 +1,17 @@
 from eosclubhouse.libquest import Quest
-from eosclubhouse.system import Sound
 
 
 class Quickstart(Quest):
 
-    __quest_name__ = 'Tutorial - The Clubhouse'
+    __quest_name__ = 'Tutorial - The Clubhouse - Quickstart'
     __tags__ = ['mission:ada']
-    __mission_order__ = 10
+    __auto_offer_info__ = {'confirm_before': False, 'start_after': 3}
+    __available_after_completing_quests__ = ['FirstContact']
+
+    def setup(self):
+        self.auto_offer = True
+        # Hide quest in UI
+        self.skippable = True
 
     def step_begin(self):
         self.wait_confirm('WELCOME1')
@@ -23,10 +28,4 @@ class Quickstart(Quest):
                 self.wait_confirm(msgid)
 
         self.wait_confirm('END1')
-        self.show_message('END2', choices=[('Got it!', self.step_end)])
-
-    def step_end(self):
-        self.complete = True
-        self.available = False
-        Sound.play('quests/quest-complete')
-        self.stop()
+        self.show_message('END2', choices=[('Got it!', self.step_complete_and_stop)])
